+++ Tests/NablaMessagingUITests/SnapshotTests/Conversation/DeletedMessageContentViewTests.swift
@@ -28,12 +28,28 @@ final class DeletedMessageContentViewTests: XCTestCase {
         assertSnapshot(matching: sut, as: .image(size: size))
     }
 
-    func testDeletedConfigureThem() {
+    func testDeletedConfigureProvider() {
         // GIVEN
         // WHEN
         sut.configure(
             with: .init(
-                sender: .them(.init(author: .authorStub, avatar: .init(url: nil, text: .initialsStub), isContiguous: false)),
+                sender: .provider(.init(author: .authorStub, avatar: .init(url: nil, text: .initialsStub), isContiguous: false)),
+                footer: nil,
+                replyTo: nil,
+                content: .init(text: "Deleted Message"),
+                menuElements: []
+            )
+        )
+        // THEN
+        assertSnapshot(matching: sut, as: .image(size: size))
+    }
+    
+    func testDeletedConfigureOther() {
+        // GIVEN
+        // WHEN
+        sut.configure(
+            with: .init(
+                sender: .other(.init(author: .otherAuthorStub, avatar: .init(url: nil, text: .otherInitialsStub), isContiguous: false)),
                 footer: nil,
                 replyTo: nil,
                 content: .init(text: "Deleted Message"),