+++ Tests/NablaMessagingUITests/SnapshotTests/Conversation/ImageMessageContentViewTests.swift
@@ -28,12 +28,28 @@ final class ImageMessageContentViewTests: XCTestCase {
         assertSnapshot(matching: sut, as: .wait(for: 0.5, on: .image(size: size)))
     }
 
-    func testImageConfigureThem() {
+    func testImageConfigureProvider() {
         // GIVEN
         // WHEN
         sut.configure(
             with: .init(
-                sender: .them(.init(author: .authorStub, avatar: .init(url: nil, text: .initialsStub), isContiguous: false)),
+                sender: .provider(.init(author: .authorStub, avatar: .init(url: nil, text: .initialsStub), isContiguous: false)),
+                footer: nil,
+                replyTo: nil,
+                content: .init(originalImageSize: .init(width: 200, height: 200), imageSource: .url(.stubImage)),
+                menuElements: []
+            )
+        )
+        // THEN
+        assertSnapshot(matching: sut, as: .wait(for: 0.5, on: .image(size: size)))
+    }
+    
+    func testImageConfigureOther() {
+        // GIVEN
+        // WHEN
+        sut.configure(
+            with: .init(
+                sender: .other(.init(author: .otherAuthorStub, avatar: .init(url: nil, text: .otherInitialsStub), isContiguous: false)),
                 footer: nil,
                 replyTo: nil,
                 content: .init(originalImageSize: .init(width: 200, height: 200), imageSource: .url(.stubImage)),